requirements.txt
flask==3.0.0
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
google-generativeai==0.3.1
Pillow==10.1.0
